website/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
itsdangerous==2.1.2
torch>=2.0.0
librosa>=0.10.0
scipy>=1.11.0
numpy>=1.24.0
matplotlib>=3.7.0
